move encode admin user component
