+++ src/app/admin/admin-routing.module.ts
@@ -4,7 +4,7 @@ import { AdminAuthGuard } from './admin-auth.guard';
 import { AdminCreativityComponent } from './components/admin-creativity/admin-creativity.component';
 import { AdminEncodeComponent } from './components/admin-encode/admin-encode.component';
 import { AdminRulitComponent } from './components/admin-rulit/admin-rulit.component';
-import { AdminEncodeUserComponent } from './components/admin-encode-user/admin-encode-user.component';
+import { AdminEncodeUserComponent } from './components/admin-encode/encode-user/admin-encode-user.component';
 
 
 import { AdminComponent } from './components/admin.component';
