+++ src/app/admin/components/admin-encode/encode-user/admin-encode-user.component.scss
