+++ src/app/admin/components/admin-encode/encode-user/admin-encode-user.component.ts
@@ -1,11 +1,11 @@
-import { ActivatedRoute, Router } from '@angular/router';
+import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { DataDbService } from 'src/app/core/services/db/data-db.service';
 
 @Component({
   selector: 'app-admin-encode-user',
   templateUrl: './admin-encode-user.component.html',
-  styleUrls: ['admin-encode-user.component.scss','../admin.component.scss'],
+  styleUrls: ['admin-encode-user.component.scss','../../admin.component.scss'],
 })
 export class AdminEncodeUserComponent implements OnInit {
 
@@ -17,11 +17,11 @@ export class AdminEncodeUserComponent implements OnInit {
     private _dbService: DataDbService,
     private route: ActivatedRoute) {}
 
-  ngOnInit(): void 
+  async ngOnInit(): Promise<void> 
   {
     let userIdParam = this.route.snapshot.paramMap.get('userId');
-    this._dbService.getEncodeUser(userIdParam).then(data => this.userData = data);
-    console.log(this.userData);
+    const userData = await this._dbService.getEncodeUser(userIdParam);
+    this.userData = userData;
   }
 
 }
